requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
faiss-cpu
openai
python-dotenv
langchain
streamlit
requests
beautifulsoup4
pinecone-client
langchain
openai
python-dotenv